Patch out insecure dependencies and add NixOS instructions
This commit uses `patchelf` to remove the dependency on OpenSSL 1.0,
which is outdated and considered insecure. It replaces this dependency
with libcurl, of which the latest version works just fine.

This should make sure that any distrobution can run the generated binary
out of the box, without needing a dynamically linked placeholder.

I also noticed the game failed to start on my AMD machine, and after looking
through Mesa's Gitlab I found a workaround using an environment variable.
A wrapper bash script that sets this variable and executes the `runner` binary
is created to take care of this automatically.

This commit adds instructions on how to run the game on NixOS as well.
I will also PR it to nixpkgs (our package repository) to make the
process a bit more seamless in the future.